Move the added cache out from the class
Oversight: Since we run in a threaded server, it's possible to
be executing multiple independent queries at the same time. Using
an instance property to track state is thus wrong.

Given the amount of state now being transmitted through the life
of a request should consider an independent object for it.