Use locking around zone and server hashes.
Ahead of future dynamic updates, we need to make operations on the
zones and servers hashed atomic.

Currently this only locks file load and initial lookups in the hash.
The reasoning here is that at the moment the details inside part
of the hash structure won't change - just adds/deletes of the primary
keys - so once we have a reference to an item in the hash, it's safe
to use outside the lock.