Swap one hack for another.
setup.py would not run if package dependencies were not yet installed
since we tried to import FlirbleDNSServer (to get version/description
values).

Instead, we parse FlirbleDNSServer/__init__.py and strip out its
imports and write that to a temporary module. We then import that
module to get at the metadata.

Naturally, we than have to catch all sorts of conditions and do
cleanup of all the files we just made (including the compiled
python bytecode, if any).

It's a hack, but it should be more robust than before. Maybe.